Fix Register64.mul; Add Register64.mod; Add support for argon2 on web platform
