+++ test/key_derivators/argon2_test.dart
@@ -1,4 +1,3 @@
-@TestOn("vm")
 import 'dart:convert';
 import 'dart:typed_data';
 
@@ -18,39 +17,44 @@ const int DEFAULT_OUTPUTLEN = 32;
 /// is a 1:1 port of BouncyCastle's Java implementation.
 void main() {
   group('Argon2BytesGenerator', () {
-    test('Hash Tests', () {
-      var version = Argon2Parameters.ARGON2_VERSION_10;
-
-      /* Multiple test cases for various input values */
+    /* Multiple test cases for various input values */
+    test('Argon2 Test 1', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_10,
           2,
           16,
           'password',
           'somesalt',
           'f6c4db4a54e2a370627aff3db6176b94a2a209a62c8e36152711802f7b30c694',
           DEFAULT_OUTPUTLEN);
-
+    });
+    // A memory-cosing test, will fail on Web platform.
+    test('Argon2 Test 2', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_10,
           2,
           20,
           'password',
           'somesalt',
           '9690ec55d28d3ed32562f2e73ea62b02b018757643a2ae6e79528459de8106e9',
           DEFAULT_OUTPUTLEN);
-
+    }, onPlatform: {
+      'chrome': Skip('Due to high memory occupation, Chrome will die.'),
+    });
+    test('Argon2 Test 3', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_10,
           2,
           16,
           'password',
           'diffsalt',
           '79a103b90fe8aef8570cb31fc8b22259778916f8336b7bdac3892569d4f1c497',
           DEFAULT_OUTPUTLEN);
+    });
 
+    test('Argon2 Test 4', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_10,
           2,
           16,
           'password',
@@ -59,39 +63,43 @@ void main() {
               '39feba4a9cd9cc5b4c798f2aaf70eb4bd044c8d148decb569870dbd923430b82a083f284beae777812cce18cdac68ee8ccef'
               'c6ec9789f30a6b5a034591f51af830f4',
           112);
-
-      version = Argon2Parameters.ARGON2_VERSION_13;
-
+    });
+    test('Argon2 Test 5', () {
       /* Multiple test cases for various input values */
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_13,
           2,
           16,
           'password',
           'somesalt',
           'c1628832147d9720c5bd1cfd61367078729f6dfb6f8fea9ff98158e0d7816ed0',
           DEFAULT_OUTPUTLEN);
-
+    });
+    test('Argon2 Test 6', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_13,
           2,
           20,
           'password',
           'somesalt',
           'd1587aca0922c3b5d6a83edab31bee3c4ebaef342ed6127a55d19b2351ad1f41',
           DEFAULT_OUTPUTLEN);
-
+    }, onPlatform: {
+      'chrome': Skip('Due to high memory occupation, Chrome will die.'),
+    });
+    test('Argon2 Test 7', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_13,
           2,
           18,
           'password',
           'somesalt',
           '296dbae80b807cdceaad44ae741b506f14db0959267b183b118f9b24229bc7cb',
           DEFAULT_OUTPUTLEN);
-
+    });
+    test('Argon2 Test 8', () {
       _hashTest(
-          version,
+          Argon2Parameters.ARGON2_VERSION_13,
           2,
           8,
           'password',
@@ -99,6 +107,10 @@ void main() {
           '89e9029f4637b295beb027056a7336c414fadd43f6b208645281cb214a56452f',
           DEFAULT_OUTPUTLEN);
     });
+  }, onPlatform: {
+    'node': Skip('Node is not supported currently, because this is quite a '
+        'high memory occupation implementation. A fatal error '
+        '\'JavaScript heap out of memory\' will be thrown.'),
   });
 }
